Fix Mac Support
Mac needs some of the things that Linux doesnt.
However rather than building a completely seperate binary
we just check the platform and call the appropriate methods.
There is still the issue of

	socket.SetIPProtectionLevel

which is NOT available on Mono. Looks like that is the only
reason we need a seperate binary for Windows.